refactor: move layout component and add navitem component
