+++ src/components/layouts/sidebar/Navbar.jsx
@@ -3,6 +3,7 @@ import { MdDashboard, MdLogout } from "react-icons/md";
 import { FaNewspaper, FaCloudSun } from "react-icons/fa6";
 import { NavLink } from "react-router";
 import { useSelector } from "react-redux";
+import NavItem from "./NavItem";
 
 const NAVBAR_ELEMENTS = {
   header: [{ tabName: "Weather App", icon: <FaCloudSun />, url: "/" }],
@@ -12,26 +13,6 @@ const NAVBAR_ELEMENTS = {
   ],
   footer: [{ tabName: "Logout", icon: <MdLogout />, url: "/" }],
 };
-
-const NavItem = ({ element, section, isSidebarOpen }) => (
-  <li className="text-white px-4 py-2 text-center">
-    <NavLink to={element.url}>
-      <div
-        className={`flex rounded-md p-2 cursor-pointer text-white text-sm items-center gap-x-4 
-        ${section === "header" ? "mb-2" : "hover:bg-white hover:text-blue-700 transition-colors delay-250 ease-out"}`}
-      >
-        {section === "header"
-          ? React.cloneElement(element.icon, { size: 32 })
-          : element.icon}
-        {isSidebarOpen && (
-          <span className={section === "header" ? "text-xl font-bold" : ""}>
-            {element.tabName}
-          </span>
-        )}
-      </div>
-    </NavLink>
-  </li>
-);
 
 const Navbar = () => {
   const isSidebarOpen = useSelector((state) => state.sidebar.isOpen);
